requirements.txt
pillow==6.1.0
pyocr==0.7.2
